requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.4.0
dnspython==1.16.0
eventlet==0.24.1
fake-useragent==0.1.11
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SocketIO==3.3.2
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
Pillow==6.0.0
pytesseract==0.2.6
python-engineio==3.5.1
python-socketio==3.1.2
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.24.1
validators==0.12.4
Werkzeug==0.15.2
